repler/requirements.txt
Twisted==16.1.1
argparse==1.2.1
attrs==15.2.0
autobahn==0.14.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==2.0.0
cffi==1.6.0
cryptography==1.3.2
docker-py==1.8.1
enum34==1.1.5
idna==2.1
ipaddress==1.0.16
pyOpenSSL==16.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
requests==2.10.0
service-identity==16.0.0
six==1.10.0
txaio==2.5.1
websocket-client==0.37.0
wsgiref==0.1.2
zope.interface==4.1.3
